Create dedicated write tests to speed up overall test time
